feat: Add tasklist command to retrieve process information by PID
This commit adds the tasklist command to retrieve process information by PID. The command is used to query the running processes on the system and filter them by a specific PID. This addition enhances the functionality of the code by providing a way to obtain detailed information about a specific process.

Note: The recent user commits and recent repository commits do not provide any relevant information for generating the commit message.