feat: Update HTTP POST form data in hydra command
This commit updates the HTTP POST form data in the hydra command for the URL "192.168.50.201". The previous form data was using the incorrect parameter name "usr" instead of "user". This change ensures that the correct parameter name is used, fixing the issue with the login functionality.

Note: The recent user commits and recent repository commits do not provide any relevant information for generating the commit message.