feat: Add gpp-decryp cmd for SMB enumeration and relay target generation
This commit adds the gpp-decryp cmd to the README.md file for SMB enumeration. The command can be used to decrypt Group Policy Preferences (GPP) passwords. Additionally, it includes a command for generating a list of relay targets. These enhancements improve the functionality of the code by providing additional tools for SMB enumeration and relay attack preparation.