feat: Update Responder.conf to disable SMB and enable SOCKS
This commit modifies the Responder.conf file to disable SMB and enable SOCKS. The sed command in the README.md file is updated to include the replacement of "HTTP = On" with "HTTP = Off" and the addition of "impacket-ntlmrelayx -smb2support -tf relayTargets.txt -socks" command. These changes enhance the security and functionality of the code by preventing SMB-related vulnerabilities and providing the ability to use SOCKS for relay attacks.